Adding level capabilities API explicitly to hashed level
