+++ include/xsparse/level_capabilities/coordinate_iterate.hpp
@@ -4,6 +4,7 @@
 #include <iterator>
 #include <tuple>
 #include <optional>
+#include <unordered_map>
 #include <type_traits>
 #include <utility>
 
@@ -11,6 +12,7 @@
 
 #include <xsparse/util/template_utils.hpp>
 #include <xsparse/util/base_traits.hpp>
+#include <xsparse/util/container_traits.hpp>
 
 
 namespace xsparse::level_capabilities
@@ -267,6 +269,99 @@ namespace xsparse::level_capabilities
         iteration_helper iter_helper(typename BaseTraits::I i, typename BaseTraits::PKM1 pkm1)
         {
             return iteration_helper{ *static_cast<typename BaseTraits::Level*>(this), i, pkm1 };
+        }
+    };
+
+    template <template <class...> class T,
+              class IK,
+              class PK,
+              class ContainerTraits,
+              class... LowerLevels>
+    class coordinate_pos_locate_iterate
+    {
+        using BaseTraits = util::base_traits<T, IK, PK, LowerLevels...>;
+
+    public:
+        class iteration_helper
+        {
+            static_assert(std::is_nothrow_invocable_r_v<std::optional<typename BaseTraits::PK>,
+                                                        decltype(&BaseTraits::Level::locate),
+                                                        typename BaseTraits::Level&,
+                                                        typename BaseTraits::PKM1,
+                                                        typename BaseTraits::IK>);
+
+        private:
+            typename ContainerTraits::template Map<typename BaseTraits::IK, typename BaseTraits::PK>
+                m_map;
+
+        public:
+            class iterator;
+            using key_type = typename BaseTraits::IK;
+            using value_type = typename std::pair<typename BaseTraits::IK, typename BaseTraits::PK>;
+            using difference_type = typename std::make_signed_t<typename BaseTraits::PK>;
+            using pointer = typename std::pair<typename BaseTraits::IK, typename BaseTraits::PK>*;
+            using reference = std::pair<typename BaseTraits::IK, typename BaseTraits::PK>;
+            using iterator_type = iterator;
+
+            class iterator : public xtl::xrandom_access_iterator_base2<iteration_helper>
+            {
+            private:
+                using wrapped_iterator_type =
+                    typename ContainerTraits::template Map<typename BaseTraits::IK,
+                                                           typename BaseTraits::PK>::const_iterator;
+                wrapped_iterator_type wrapped_it;
+
+            public:
+                explicit inline iterator(wrapped_iterator_type wrapped) noexcept
+                    : wrapped_it(wrapped)
+                {
+                }
+
+                inline std::tuple<typename BaseTraits::IK, typename BaseTraits::PK> operator*()
+                    const noexcept
+                {
+                    return { wrapped_it->first, wrapped_it->second };
+                }
+
+                inline bool operator==(const iterator& other) const noexcept
+                {
+                    return wrapped_it == other.wrapped_it;
+                }
+
+                inline iterator& operator++() noexcept
+                {
+                    ++wrapped_it;
+                    return *this;
+                }
+
+                inline iterator& operator--() noexcept
+                {
+                    --wrapped_it;
+                    return *this;
+                }
+            };
+
+            explicit inline iteration_helper(
+                typename ContainerTraits::template Map<typename BaseTraits::IK,
+                                                       typename BaseTraits::PK>& map) noexcept
+                : m_map(map)
+            {
+            }
+
+            inline iterator_type begin() const noexcept
+            {
+                return iterator_type{ m_map.begin() };
+            }
+
+            inline iterator_type end() const noexcept
+            {
+                return iterator_type{ m_map.end() };
+            }
+        };
+
+        iteration_helper iter_helper(typename BaseTraits::PKM1 pkm1)
+        {
+            return iteration_helper{ this->m_crd[pkm1] };
         }
     };
 }
